Create Unstop Daily challenge 1
One day Mary wanted to give a present to her friend, and she decided to give her a beautiful bouquet of flowers, she started collecting the flowers for the bouquet, she exactly needed 2 types of flowers. The total number of flowers should be ‘t’. Now she starts collecting it from her garden. In her garden, there are ‘N’ types of flowers, and each type of flower are arranged in a queue in non-decreasing order, e.g. 1, 3, 6, 15 and so on.

Now she wants your help to find which indexes of flowers she should collect.